add no delete file source

+++ src/traits/Upload.php
@@ -3,15 +3,22 @@
 namespace fractalCms\traits;
 
 use Exception;
-use fractalCms\models\ElasticModel;
-use fractalCms\Module;
 use Yii;
-use yii\helpers\ArrayHelper;
 
 trait Upload
 {
 
-    public function saveFile($dataFile, $relativeDirName, $value) : string
+    /**
+     * Save file in webroot
+     *
+     * @param $dataFile
+     * @param $relativeDirName
+     * @param $value
+     * @param $deleteSource
+     * @return string
+     * @throws Exception
+     */
+    public function saveFile($dataFile, $relativeDirName, $value, $deleteSource = true) : string
     {
         try {
             $result = $value;
@@ -32,7 +39,9 @@ trait Upload
                         $newPath = Yii::getAlias($dataFile.'/'.$relativeDirName.'/'.$this->id.'/'.$fileName);
                         $success = copy($filePatch, $newPath);
                         if ($success === true) {
-                            unlink($filePatch);
+                            if ($deleteSource === true) {
+                                unlink($filePatch);
+                            }
                             $result = $dataFile.'/'.$relativeDirName.'/'.$this->id.'/'.$fileName;
                         }
                     }
